twitter/functions/twitter_ingest_primary_get_timeline/requirements.txt
google-cloud-secret-manager==2.2.0
google-cloud-firestore==2.0.2
google-cloud-pubsub==1.5.0
elasticsearch>=7.0.0,<8.0.0
tweepy==3.8.0
azure-cosmosdb-table==1.0.5
